examples/research/requirements.txt
-e ../../
langgraph-cli[inmem]
tavily-python
python-dotenv
